added geo tracking

+++ src/main.ts
@@ -44,23 +44,38 @@ document.getElementById("north")?.addEventListener("click", () => {
   player.position.lat += board.tileWidth;
   player.marker.setLatLng(player.position);
   board.drawPits(player.position, map);
+  map.setView(player.position);
 });
 document.getElementById("south")?.addEventListener("click", () => {
   player.position.lat -= board.tileWidth;
   player.marker.setLatLng(player.position);
   board.drawPits(player.position, map);
+  map.setView(player.position);
 });
 document.getElementById("east")?.addEventListener("click", () => {
   player.position.lng += board.tileWidth;
   player.marker.setLatLng(player.position);
   board.drawPits(player.position, map);
+  map.setView(player.position);
 });
 document.getElementById("west")?.addEventListener("click", () => {
   player.position.lng -= board.tileWidth;
   player.marker.setLatLng(player.position);
   board.drawPits(player.position, map);
+  map.setView(player.position);
 });
 document.getElementById("reset")?.addEventListener("click", () => {
+  if (window.prompt("Do you want to reset? (yes/no)") !== "yes") return;
   localStorage.clear();
   window.location.reload();
+});
+
+const sensorButton = document.querySelector("#sensor")!;
+sensorButton.addEventListener("click", () => {
+    navigator.geolocation.watchPosition((position) => {
+      player.position = leaflet.latLng(position.coords.latitude, position.coords.longitude);
+      player.marker.setLatLng(player.position);
+      map.setView(player.position);
+      board.drawPits(player.position, map);
+    });
 });
